Replaces ConcurrentRdaSink with ReactiveRdaSink.
