+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc/src/test/java/gov/cms/bfd/pipeline/rda/grpc/sink/concurrent/SequenceNumberWriterTest.java
@@ -1,4 +1,4 @@
-package gov.cms.bfd.pipeline.rda.grpc.sink.reactive;
+package gov.cms.bfd.pipeline.rda.grpc.sink.concurrent;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -8,7 +8,6 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 
 import gov.cms.bfd.pipeline.rda.grpc.RdaSink;
-import gov.cms.bfd.pipeline.rda.grpc.sink.concurrent.SequenceNumberTracker;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -22,7 +21,10 @@ public class SequenceNumberWriterTest {
   /** Mock used for verifying database writes. */
   @Mock private RdaSink<Integer, Long> sink;
 
-  /** Instance of {@link SequenceNumberTracker} passed to the {@link SequenceNumberWriter}. */
+  /**
+   * Instance of {@link gov.cms.bfd.pipeline.rda.grpc.sink.concurrent.SequenceNumberTracker} passed
+   * to the {@link SequenceNumberWriter}.
+   */
   private SequenceNumberTracker tracker;
 
   /** Instance being tested. */
